feat(SPV-737): search contacts
