+++ actions/common/search_methods.go
@@ -3,20 +3,12 @@ package common
 import (
 	"math"
 
-	"github.com/bitcoin-sv/spv-wallet/engine/datastore"
 	"github.com/bitcoin-sv/spv-wallet/models"
+	"github.com/bitcoin-sv/spv-wallet/models/filter"
 )
 
-// LoadDefaultQueryParams will load the default query parameters
-func LoadDefaultQueryParams() *datastore.QueryParams {
-	return &datastore.QueryParams{
-		Page:     1,
-		PageSize: 10,
-	}
-}
-
 // GetPageFromQueryParams will return a Page object from the query parameters and count value
-func GetPageFromQueryParams(queryParams *datastore.QueryParams, count int64) models.Page {
+func GetPageFromQueryParams(queryParams *filter.QueryParams, count int64) models.Page {
 	totalPages := int(math.Ceil(float64(count) / float64(queryParams.PageSize)))
 	page := models.Page{
 		Size:          queryParams.PageSize,
