+++ actions/contacts/search.go
@@ -6,6 +6,7 @@ import (
 	"github.com/bitcoin-sv/spv-wallet/actions/common"
 	"github.com/bitcoin-sv/spv-wallet/mappings"
 	"github.com/bitcoin-sv/spv-wallet/models"
+	"github.com/bitcoin-sv/spv-wallet/models/filter"
 	"github.com/bitcoin-sv/spv-wallet/server/auth"
 	"github.com/gin-gonic/gin"
 )
@@ -25,7 +26,7 @@ import (
 func (a *Action) search(c *gin.Context) {
 	reqXPubID := c.GetString(auth.ParamXPubHashKey)
 
-	var reqParams SearchContacts
+	var reqParams filter.SearchContacts
 	if err := c.Bind(&reqParams); err != nil {
 		c.JSON(http.StatusBadRequest, err.Error())
 		return
@@ -37,16 +38,14 @@ func (a *Action) search(c *gin.Context) {
 		return
 	}
 
-	if reqParams.QueryParams == nil {
-		reqParams.QueryParams = common.LoadDefaultQueryParams()
-	}
+	reqParams.DefaultsIfNil()
 
 	contacts, err := a.Services.SpvWalletEngine.GetContactsByXpubID(
 		c.Request.Context(),
 		reqXPubID,
-		reqParams.Metadata,
+		mappings.MapToMetadata(reqParams.Metadata),
 		conditions,
-		reqParams.QueryParams,
+		mappings.MapToQueryParams(reqParams.QueryParams),
 	)
 	if err != nil {
 		c.JSON(http.StatusInternalServerError, err.Error())
@@ -58,7 +57,7 @@ func (a *Action) search(c *gin.Context) {
 	count, err := a.Services.SpvWalletEngine.GetContactsByXPubIDCount(
 		c.Request.Context(),
 		reqXPubID,
-		reqParams.Metadata,
+		mappings.MapToMetadata(reqParams.Metadata),
 		conditions,
 	)
 	if err != nil {
